Simplified code + other fixes
Unnecessary bootstrap has been removed throughout the site. Also, the nomenclature of the html has been improved. Dashes are used consistently in id & className as opposed to camel case. All scss & app.js references to these ids & classes were changed accordingly. A problem with the navbar changing back to gray near the bottom of the page was fixed. Clicking on the navbar elements now takes to you the top of the corresponding div instead of the top of the header. This is more natural, as now there is spacing between the top of the viewport and the headers. #goToTop was changed to #thanks to match the text change. The text in the tab was also changed to 'Homepage'. Animations were added to the icons in Hobbies.js & take place when the element hits the bottom of the screen. Finally, a tooltip was added to the arrow at the top of the screen. More will be added in the next update.