Video update & more
Video has been added to the site. The videos sit inside both windows on the page & run on a loop. The Webpack file had to be altered to allow multiple videos to be added to the site. Also, an icon had been added to the rightmost area of the navbar. As the user scrolls through different sections of the site, the icon changes its icon. More unnecessary bootstrap has been removed from the page, & now the user maneuver through the page with only a keyboard. Inside Education-2, there were visibility problems with the list items, so the background was darkened & the text was lighened. Lasly, naming inconsistencies were addressed & odd coding structure is being revamped.