New Icons, divs, videos, & more
New menu & close icons were added. These icons have are circular, which better matches the site's design. There are no more glyphicons used on the page. A new Employment.js file was added, which consists of two divs (.Employment-1 & .Employment-2). .Employment-1 follow the design of .Intro-1, which a large circular video at the top along with a header & paragraph below. Also, a div was added in Education.js which shows off my senior project. This div shows off a new div design with a header & rows of circles inside a flex-container. Five of the circles contain images of the drone while the top center circle contains a flight video. On hover, each circle darkens & acid green text appears. This design will eventually be used for Hobbies.js. As for minor fixes, about-me-paragraph now has smoother fade-in animations, camel case naming is being replaced with bootstrap dash (-) notation, a problem that caused the menu to move down on click was fixed, & some header/li text was changed. Lastly, the menu icon moves up slightly when the navbar is pinned to the top of the viewport.