Massive Update
Many cross-browser problems have been fixed. Brightness Tweens on Intro-1 were removed in Firefox due to lag. Background shaking in Languages-2 using Firefox was fixed by setting the background-attachment to fixed. Background image proportion issues were fixed in Microsoft Edge by setting the background-size in the affected divs to fixed. Background shaking in Volunteer-2 in Edge was fixed by setting the background-attachment to fixed. All headers on the page were changed to h1s. The borders on #employment now change to acid green on hover, matching the behavior of the rest of the site. Formerly, circular flex-items had an onHover animation that took place when the cursor was outside of the area. This has been fixed. Extra Hobbies-2 & Volunteer-3 divs were added. Hobbies-1 & the Sidebar have been revamped to match the format of the rest of the site. Finally, the src of #down-arrow was changed so that an invert filter no longer has to be used.