Fixing browser/performance issues & more
Microsoft Edge has problems rendering circular videos, so specifically for Edge, all circular flex-items are square. Edge was also having problems with the scroll animation for .Hobbies-4, so it was not applied in that browser. Firefox issues were also corrected. Firefox has issues with the scroll animations in .Education-1 & .Languages-2, so they aren't applied in the browser. For mobile, the most demanding parallax animations were removed to boost performance. The pushMenuUpTween has been removed & #menu now maintains a fixed position throughout the entire page. Extra margin has been added to all headers with a corresponding #navbar link. This makes it so the header isn't cut off by the navbar after jumping to a certain section of the page. Other small changes include removing unnecessary font imports, adding a space between the words in #name, decreasing #name's font-size on the for-phone-landscape breakpoint, implementing more consistent nomenclature for the videos & changing smaller contentless divs from 40vh to 35vh on mobile. Futhermore, on mobile phones in portrait mode, the text in the flex-items was too small. To fix this, circular flex-items have been enlarged to 600px & the text inside has also been enlarged to 4em for the headers & 1.7em for the pagagraphs. This change puts only puts one flex-item on each row as opposed to two previously. Lastly, #nav-icon is now a div & the background-image changes as the user scrolls through the sections. This creates a smoother icon change from secction to section.