New .Footer-1 markup, faster loading, & more
Flex boxes are now used in .Footer-1 as opposed to Bootstrap. Peformance improvements include shrinking & compressing the sidebar logo images, asynchronously loading jQuery & Bootstrap in the index.html file, & skimming some bytes off several backgroud images. #name's border no longer changes to -green on hover. This was removed because the brighting on .Intro-1 that takes place as the user scrolls down changes the green to yellow hue, which doesn't match the color scheme of the page. Class exporting is now done all on one line in all component files. Lastly, the background-position of .Skills-1 was shifted to the right & #down-arrow was moved up slightly on mobile.