Cropped images & deleted unnecessary scss
Many images used for the circular flexboxes were rectangular. Therefore, only ~40% of the entire rectangular image was actually visible to the user. All of these images are now perfect squares, which lowers the percentage of unseen part of the image. In total, this cut over 500KB off the total file size of the images. Also, some unnecessary scss was removed.