Microsoft Edge circular video issue fixed
Microsoft Edge doesn't render circular videos properly. Circular videos were made on this page by either using a circular clip-path or a 50% border-radius on the parent div. In Microsoft Edge however, a rectangular video is generated with all four edges aligning to the edge of the circle. To fix this, code was added that removes videos only in Microsoft Edge, & new images are added in their place. These new images are all shrunken & compressed.