Improved sidebar + removed divs
The sidebar now has a dropdown menu that can be toggled onClick. The menu is open by default & has links to all the individual GitHub projects. When hovered over, a preview of each page is shown to the right of the sidebar (the hover effect only takes place on larger screens however). The pexels link has been removed from the sidebar as it was unnecessary. Also, .skills-div-1 now has a simple -gray background & one of the content-less divs at the bottom of the page has been removed. Both of these were done to reduce the load time of the page. Future performance improvements will be coming soon.