New design + faster load times
Three out of the four videos were removed to speed up load times. Additionally, a new background pattern was added to all divs that formerly had a solid -gray background. Instead of adding the background directly, the background pattern is applied to the app's background & the aforementioned divs have a transparent background. Lastly, the quadrocopter video is now visible on mobile & a shadow has been added to the sidebar when the openSidebar() state is applied.